feat(shell): add tests for validate-address
