+++ libs/shared/src/utils/convert-address.ts
@@ -20,17 +20,3 @@ export function ethToHaqq(address: string) {
 export function haqqToEth(address: string) {
   return bech32Decode(address);
 }
-
-export function isValidHaqqAddress(address: string) {
-  try {
-    bech32Decode(address);
-
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-export function isValidEthAddress(address: string) {
-  return /^0x[a-fA-F0-9]{40}$/.test(address);
-}
